src/requirements.txt
requests
flask
Flask
python-dotenv